out/production/lhrkits/requirements/requirements.txt
--index http://pypi.mirrors.ustc.edu.cn/simple
# --index http://mirrors.ustc.edu.cn/pypi/web/simple/
docopt
sh
pyyaml
biopython
numpy
mongokit
pymongo==2.8   ## 这里必须用2.8, 否则和mongokit不匹配, 会出错
rq
pandas
clize
requests



flask==0.10.1
flask-markdown
Flask-Mail
gunicorn
python-docx==0.7.6
lxml==3.4.0
